Remove extra newline from before list <pre> blocks
My crowded-pre fix broke <pre> blocks in lists:

  $ diff -u preformatted_in_list.md preformatted_in_list-module_output.md
  --- preformatted_in_list.md
  +++ preformatted_in_list-module_output.md
  @@ -1,5 +1,6 @@
     * Run this command:

  +
           ls -l *.html

     * ?

There is a fair amount of trailing whitespace in html2text output, and
I'm not sure where it all comes from.  This patch removes the extra
newline (fixing the test), but it also tweaks the amount of trailing
whitespace in the expected blank line (probably not a problem).